refactored pip install and implemented deploy_project. Uses rsync with existing version to minimize transfer. Pip install no longer allows pip_requirements setting

+++ woven/virtualenv.py
@@ -13,7 +13,11 @@ from fabric.contrib.files import exists
 from fabric.contrib.project import rsync_project, upload_project
 from fabric.contrib.console import confirm
 
+#Required for a bug in 0.9
+from fabric.version import get_version
+
 from woven.utils import root_domain, server_state, set_server_state
+from woven.utils import project_fullname, project_name, active_version
 
 class Virtualenv(object):
     """
@@ -22,13 +26,14 @@ class Virtualenv(object):
     """
     state = 'created_virtualenv_'
     def __init__(self, version=''):
-        if not version: self.version = env.project_fullname
-        else: self.version = env.project_name+'-'+version
+        if not version: self.fullname = project_fullname()
+        else: self.fullname = project_name()+'-'+version
+
         env.deployment_root = '/home/%s/%s/'% (env.user,root_domain())
         self.root = env.deployment_root +'env' 
-        self.path = '/'.join([self.root,self.version])
+        self.path = '/'.join([self.root,self.fullname])
         self.python_path = '/'.join([self.path,'bin','python'])
-        if server_state(self.state+self.version):self.installed = True
+        if server_state(self.state+self.fullname):self.installed = True
         else: self.installed = False
 
 def mkvirtualenv(version=''):
@@ -38,7 +43,7 @@ def mkvirtualenv(version=''):
     v = Virtualenv(version)
     if v.installed:
         if env.verbosity:
-            print env.host,'Warning: Virtualenv %s already installed. Skipping..'% v.version
+            print env.host,'Warning: Virtualenv %s already installed. Skipping..'% v.fullname
         return False
     else:
         #TODO: We need the apache conf to use WSGIPythonHome /usr/local/pythonenv/BASELINE
@@ -53,9 +58,9 @@ def mkvirtualenv(version=''):
         run('mkdir -p '+v.root)
         with cd(v.root):
             if no_site:
-                run("virtualenv --no-site-packages %s" % v.version)
+                run("virtualenv --no-site-packages %s" % v.fullname)
             else:
-                run("virtualenv %s" % v.version)
+                run("virtualenv %s" % v.fullname)
 
         # some installations require an egg-cache that is writeable
         # by the apache user - normally www-data
@@ -66,7 +71,7 @@ def mkvirtualenv(version=''):
             sudo('chmod -R g+w egg_cache')
 
         #Set the state to virtualenv created
-        set_server_state('created_virtualenv_'+v.version)
+        set_server_state('created_virtualenv_'+v.fullname)
         return True
 
 def rmvirtualenv(version=''):
@@ -76,7 +81,7 @@ def rmvirtualenv(version=''):
     v = Virtualenv(version)
     if v.installed: #delete
         sudo('rm -rf '+v.path)
-        set_server_state('created_virtualenv_'+v.version,delete=True)
+        set_server_state('created_virtualenv_'+v.fullname,delete=True)
     #If there are no further remaining envs we'll delete the home directory to effectively teardown the project
     if not server_state('created_virtualenv_',prefix=True):
         sudo('rm -rf '+env.deployment_root)
@@ -90,44 +95,34 @@ class Pip(Virtualenv):
         #Since pip installation uses Virtualenv we reuse existing Virtualenv functionality
         super(Pip,self).__init__(version)
         #We will store any pip cached packages and pybundles here
-        self.cache = env.deployment_root+'dist'
+        self.dist = env.deployment_root+'dist/'
+        self.cache = env.deployment_root+'package-cache/'
         #Errors get stored here (from pip defaults)
         self.pip_log = '/home/%s/.pip/pip.log'% env.user
-        #Test for pybundle existance
-        if os.path.exists('%s.pybundle'% env.project_name) and not env.PIP_REQUIREMENTS:
-            self.pip_files = ['%s.pybundle'% env.project_name]
-        elif os.path.exists('dist/%s.pybundle'% env.project_name) and not env.PIP_REQUIREMENTS:
-            self.pip_files = ['dist/%s.pybundle'% env.project_name]
-        elif not env.PIP_REQUIREMENTS:
-            #Use any file starting with req and ending with .txt
-            self.pip_files = local('ls req*.txt').rstrip().split('\n')
-        else:
-            self.pip_files = env.PIP_REQUIREMENTS
-        self.dest_files = {}
-        for file in self.pip_files:
-            dest_file = file
-            if '.pybundle' in file:
-                self.pybundle = dest_file = self.cache + '/' + file.split('/')[-1]
-            else:
-                dest_file = self.path + '/' + file
-            self.dest_files[file]=dest_file    
+        self.pip_files = []
+        self.pybundle = ''
+        #Test for pybundle existance - this must always happen first
+        #cwd = os.getcwd()
+        if os.path.exists('dist/%s.pybundle'% self.fullname):
+            #dist is the default dist-dir=DIR in distribute etc
+            self.pip_files = ['%s.pybundle'% self.fullname]
+            self.pybundle = '%s.pybundle'% self.fullname
+        #Use any file starting with req and ending with .txt
+        #Even if pybundles exist we still need requirements for uninstallation
+        self.pip_files += local('ls req*.txt').rstrip().split('\n')
             
 def pip_install_requirements(rollback=False):
     """
-    Install on current installed virtualenv version from a [project name].pybundle or pip ``req.txt``|``requirements.txt``
+    Install on current installed virtualenv version from a [dist/project name-version].pybundle or pip ``req.txt``|``requirements.txt``
     or a env.pip_requirements list.
     
-    By default it will look for a pybundle in the setup.py directory or dist directory first then a requirements file.
-    Alternatively a pybundle path can be defined in settings.PIP_REQUIREMENTS
-    
-    e.g. PIP_REQUIREMENTS = ['req1.txt','requirements.txt']
-    
-    or
-    
-    PIP_REQUIREMENTS =  ['somebundle.pybundle',...]
-    
+    By default it will look for a pybundle in the dist directory first then a requirements file.
+
     Leaves an install pip-log.txt dropping in the project dir on error which we will pick up
     and download.
+    
+    The limitations of installing requirements are that you cannot point directly to packages
+    in your local filesystem. In this case you would bundle instead.
     """
 
     p = Pip()
@@ -141,7 +136,7 @@ def pip_install_requirements(rollback=False):
             print env.host,'Pip requirements not installed. Skipping...'
         return
     #TODO - this is probably better handled in the Virtualenv init
-    elif not server_state('created_virtualenv_'+p.version):
+    elif not server_state('created_virtualenv_'+p.fullname):
         print env.host,'Error: Cannot run pip_install_requirements. A virtualenv is not created for this version. Run mkvirtualenv first'
         return False
     if env.verbosity:
@@ -152,19 +147,34 @@ def pip_install_requirements(rollback=False):
     #Uploade the req and pybundle files
     if not rollback: #create p.cache directory
         run('mkdir -p '+p.cache)
-    for file in p.dest_files:
-        put(file,p.dest_files[file])
-            
+        run('mkdir -p '+p.dist)
+    #Work around the rsync issue in 0.9
+    fab_vers = int(get_version(form='short')[0])
+    if fab_vers < 1: extra_opts = '--rsh="ssh -p%s"'% env.port
+    else: extra_opts = ''
+    exclude = ['.*']
+    #Optimize network copy
+    if p.pybundle:
+        local('mkdir -p /tmp/dist')
+        local('cp -f %s %s'% ('dist/'+p.pybundle,'/tmp/dist/%s.pybundle'% project_name()))
+        #Rsync the dist directory first 
+        rsync_project(local_dir='/tmp/dist',remote_dir=env.deployment_root,extra_opts=extra_opts,exclude=exclude,delete=False)
+        run('cp -f %s %s'% (p.dist+project_name()+'.pybundle',p.dist+p.pybundle))   
+    #put any remaining files ie requirements files up
+    for file in p.pip_files:
+        if '.pybundle' not in file:
+            put(file,p.dist+file)
     with cd(p.path):
         for req in p.pip_files:
             with settings(warn_only=True):
                 if rollback: command = 'uninstall'
-                else: command = 'install' 
-                if '.pybundle' in req:
-                    install = run('pip %s %s -q --environment=%s'% (command,p.pybundle, p.python_path))
-                else:
-                    if rollback: install = run('pip uninstall -qy --environment=%s --requirement=%s'% (p.python_path,req))
-                    else: install = run('pip install -q --environment=%s --download-cache=%s --requirement=%s'% (p.python_path,p.cache,req))
+                else: command = 'install'
+                if req == p.pybundle and not rollback:
+                    install = run('pip install %s -q --environment=%s'% (p.dist+p.pybundle, p.python_path))
+                    break #do not install anything else once a bundle is installed
+                elif req <> p.pybundle:
+                    if rollback: install = run('pip uninstall -qy --environment=%s --requirement=%s'% (p.python_path,p.dist+req))
+                    else: install = run('pip install -q --environment=%s --download-cache=%s --requirement=%s'% (p.python_path,p.cache,p.dist+req))
 
     if exists(p.pip_log) or install.failed and not rollback:
         print 'PIP errors on %s please review the pip-log.txt which will be downloaded to'% command
@@ -174,21 +184,84 @@ def pip_install_requirements(rollback=False):
     else:
         if rollback: delete = True
         else: delete = False
-        set_server_state('pip_installed_'+p.version,delete=delete)
+        set_server_state('pip_installed_'+p.fullname,delete=delete)
         if env.verbosity:
             print env.host,'PIP %sED '% command.upper(),' '.join(p.pip_files)
-    if rollback: #finally for rollback delete cache to be complete
+    if rollback: #finally for rollback and delete cache to be complete
+        for req in p.pip_files:
+            run('rm -f '+p.dist+req)
         run('rm -rf '+p.cache)
+        ls = run('ls '+ p.dist).rstrip().split('\n')
+        #If all we're left with is the generic pybundle or nothing then delete
+        if not ls[0] or ls[0]==project_name()+'.pybundle':
+            run('rm -rf '+p.dist)
         
     return True
 
 #DEPLOY
 
-class Deploy(Virtualenv):
+class Project(Virtualenv):
     """
     Base class for deploying your project
     """
+    state = 'deployed_project_'
     def __init__(self,version=''):
-        super(Deploy,self).__init__(version)
+        super(Project,self).__init__(version)
+        self.deploy_type = self.__class__.__name__.lower()
+        self.deploy_root = '/'.join([env.deployment_root,'env',self.fullname,self.deploy_type,''])
+    
+    def deploy(self,patch=''):
+        """
+        Deploy your project
+        """
+        if self.installed and not patch:
+            if env.verbosity:
+                print env.host,"Warning: %s version %s already deployed. Skipping..."% (self.deploy_type,self.fullname)
+            return False
+        elif not self.installed and not patch:
+            run('mkdir -p '+self.deploy_root)
+        elif not self.installed and patch:
+            if env.verbosity:
+                print env.host,"Warning: Cannot patch %s. This version %s does not exist. Skipping.."% (self.deploy_type,self.fullname)
+            return False
+        #bug in fabric 0.9 on rsync on alternate port fixed in 1.0
+        fab_vers = int(get_version(form='short')[0])
+        if fab_vers < 1:
+            extra_opts = '--rsh="ssh -p%s"'% env.port
 
+        exclude = ['*.pyc','.*','/build','/dist','/media','/appmedia','/www','/public']
+        #to save a bit of bandwidth if there is an existing version we will copy that first
+        current_version = active_version()
+        if current_version and not patch:
+            delete=True
+            existing_root = '/'.join([env.deployment_root+current_version,self.deploy_type])
+            run('cp -R %s %s'% (existing_root,self.deploy_root))
+        else:
+            delete=False
+        rsync_project(local_dir='./',remote_dir=self.deploy_root,extra_opts=extra_opts,exclude=exclude,delete=delete)
+        set_server_state('deployed_project_'+self.fullname)
+        if env.verbosity:
+            print "DEPLOYED %s %s"% (self.deploy_type,self.fullname)
+        return True
+    
+    def delete(self):
+        if active_version <> self.fullname:
+            run('rm -rf '+self.deploy_root)
+        #if nothing left in the specific env
+        ls = run('ls '+ self.path).rstrip().split('\n')
+        if not ls[0]: run('rm -rf '+ self.path)
+        #if no envs left
+        ls = run('ls '+ self.root).rstrip().split('\n')
+        if not ls[0]: run('rm -rf '+ env.deployment_root)
 
+     
+def deploy_project(patch=''):
+    p = Project(patch)
+    #Create an sqlite database directory if necessary
+    if env.DEFAULT_DATABASE_ENGINE == 'django.db.backends.sqlite3' and not patch:
+        db_dir = os.path.join(env.deployment_root,'database')
+        if not exists(db_dir):
+            run("mkdir -p %s"% db_dir)
+            sudo("chown %s:www-data %s"% (env.user,db_dir))
+            sudo("chmod ug+w %s"% db_dir)
+    return p.deploy(patch)
